高价函数  柯里化函数

+++ fpinjava-parent/fpinjava-usingfunctions-exercises/src/test/java/com/fpinjava/functions/exercise02_05/FunctionTest.java
@@ -1,17 +1,17 @@
 package com.fpinjava.functions.exercise02_05;
 
-import static org.junit.Assert.*;
-
 import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
 
 public class FunctionTest {
 
-  public static final Function<Integer, Integer> triple = x -> x * 3;
+  public static final Function<Double, Long> triple = x -> ((Double)(x * 3)).longValue();
 
-  public static final Function<Integer, Integer> square = x -> x * x;
+  public static final Function<Long, Integer> square = x -> Math.toIntExact(x * x);
 
   @Test
   public void test() {
-    assertEquals(Integer.valueOf(36), Function.<Integer, Integer, Integer>higherCompose().apply(square).apply(triple).apply(2));
+    assertEquals(Integer.valueOf(36), Function.<Double, Long, Integer>higherCompose().apply(square).apply(triple).apply(2d));
   }
 }
